Python RRT*-Connect Algorithm
I was not able to add a folder for the python files that I want to add, so I am adding them to the main segment, sorry about that! The RRTStarConnect is the class to run the RRT* algorithm. The RRTStarConnectTest is the python file that demonstrates how to use the RRTStarConnect algorithm.